pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "config-stash"
version = "0.0.1"
description = 'A flexible configuration management library for Python applications'
readme = "README.md"
requires-python = ">=3.9"
license = { file = "LICENSE.txt" }
keywords = ["configuration", "management", "python", "yaml", "json", "toml", "env"]
authors = [
  { name = "qatoolist", email = "contact@example.com" },
]
maintainers = [
  { name = "qatoolist", email = "contact@example.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
  "Operating System :: OS Independent",
  "Typing :: Typed",
]
dependencies = [
  "pyyaml>=6.0",
  "tomli>=2.0.0; python_version < '3.11'",
]

[project.optional-dependencies]
watch = [
  "watchdog>=2.0.0",
]
cli = [
  "click>=8.0.0",
]
dev = [
  "pytest>=7.0.0",
  "pytest-cov>=4.0.0",
  "pytest-mock>=3.10.0",
  "pytest-xdist>=3.0.0",
  "black>=23.0.0",
  "isort>=5.12.0",
  "ruff>=0.1.0",
  "mypy>=1.0.0",
  "pre-commit>=3.0.0",
  "types-toml>=0.10.0",
  "types-PyYAML>=6.0.0",
  "types-click>=8.0.0",
  "bandit>=1.7.0",
  "safety>=2.3.0",
  "radon>=6.0.0",
  "ipython>=8.0.0",
  "bump2version>=1.0.0",
]
test = [
  "pytest>=7.0.0",
  "pytest-cov>=4.0.0",
  "pytest-mock>=3.10.0",
  "pytest-xdist>=3.0.0",
]
docs = [
  "sphinx>=6.0.0",
  "sphinx-rtd-theme>=1.2.0",
  "sphinx-autodoc-typehints>=1.22.0",
]
cloud = [
  "boto3>=1.26.0",  # AWS S3
  "azure-storage-blob>=12.14.0",  # Azure Blob Storage
  "azure-identity>=1.12.0",  # Azure authentication
  "google-cloud-storage>=2.7.0",  # Google Cloud Storage
  "ibm-cos-sdk>=2.13.0",  # IBM Cloud Object Storage
  "requests>=2.28.0",  # HTTP remote loading
]
validation = [
  "pydantic>=2.0.0",  # Pydantic model validation
  "jsonschema>=4.17.0",  # JSON Schema validation
]
secrets = [
  "boto3>=1.26.0",  # AWS Secrets Manager
  "hvac>=1.1.0",  # HashiCorp Vault
  "azure-keyvault-secrets>=4.6.0",  # Azure Key Vault
  "azure-identity>=1.12.0",  # Azure authentication
  "google-cloud-secret-manager>=2.16.0",  # GCP Secret Manager
]
all = [
  "pyyaml>=6.0",
  "tomli>=2.0.0; python_version < '3.11'",
  "watchdog>=2.0.0",
  "click>=8.0.0",
  "pydantic>=2.0.0",
  "jsonschema>=4.17.0",
  "boto3>=1.26.0",
  "azure-storage-blob>=12.14.0",
  "azure-identity>=1.12.0",
  "azure-keyvault-secrets>=4.6.0",
  "google-cloud-storage>=2.7.0",
  "google-cloud-secret-manager>=2.16.0",
  "ibm-cos-sdk>=2.13.0",
  "requests>=2.28.0",
  "hvac>=1.1.0",
]

[project.urls]
Documentation = "https://github.com/qatoolist/config-stash#readme"
Issues = "https://github.com/qatoolist/config-stash/issues"
Source = "https://github.com/qatoolist/config-stash"
Changelog = "https://github.com/qatoolist/config-stash/blob/main/CHANGELOG.md"

[project.scripts]
config-stash = "config_stash.cli:cli"

# Version is now statically defined in [project] section

[tool.hatch.envs.types]
extra-dependencies = [
  "pre-commit",
  "pre-commit-hooks",
  "mypy>=1.0.0",
  "pytest",
  "ruff",
  "black",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/config_stash tests}"

[tool.mypy]
python_version = "3.9"
ignore_missing_imports = true
strict = false
warn_return_any = false
warn_unused_configs = true
no_implicit_reexport = false
warn_redundant_casts = true
warn_unused_ignores = true

[tool.coverage.run]
source_pkgs = ["config_stash", "tests"]
branch = true
parallel = true
omit = [
  "src/config_stash/__about__.py",
]

[tool.coverage.paths]
config_stash = ["src/config_stash", "*/config-stash/src/config_stash"]
tests = ["tests", "*/config-stash/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.ruff]
line-length = 88
select = ["E", "F", "W", "C", "N", "Q"]
ignore = ["E501", "E731", "E741", "F401", "F841", "N806", "C901"]

[tool.config_stash]
default_environment = "development"
default_files = ["config.yaml", "config.json", "config.toml"]
default_prefix = "PREFIX"
dynamic_reloading = false

[tool.config_stash.loaders]
yaml = "config_stash.loaders.yaml_loader:YamlLoader"
json = "config_stash.loaders.json_loader:JsonLoader"
toml = "config_stash.loaders.toml_loader:TomlLoader"
env = "config_stash.loaders.environment_loader:EnvironmentLoader"
